# BonusServer_Node/openapi.yaml
...
openapi: 3.0.0
info:
  title: 您的 API 標題
  version: "1.0.0"
  description: API 說明
servers:
  - url: http://127.0.0.1:8000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /test/login:
    post:
      summary: User login
      description: Logs in a user and returns their IP address
      tags: [Test]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                machineName:
                  type: string
                  description: The name of the machine
                scoreScale:
                  type: number
                  description: The score scale
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: login token
                  webSite:
                    type: string
                    description: name of bonus server
                  ruleId:
                    type: number
                    description: using which rule
        500:
          description: Server error
  /test/spin:
    post:
      summary: Spin the machine
      description: Spins the machine and returns the result
      tags: [Test]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                TotalBet:
                  type: number
                  description: The total bet amount
                TotalWin:
                  type: number
                  description: The total win amount
                WinA:
                  type: number
                  description: The win amount for category A
                WinB:
                  type: number
                  description: The win amount for category B
      responses:
        200:
          description: Successful spin
        500:
          description: Server error
      security:
        - BearerAuth: []   # Specify the Bearer token security scheme here
  /test/settings:
    get:
      summary: Get settings
      description: Retrieves the settings for the machine
      tags: [Test]
      responses:
        200:
          description: Successful retrieval
        500:
          description: Server error
...
  /configSetting/settings:
    get:
      summary: Get configuration settings
      description: Retrieves the current configuration settings
      tags: [Config]
      responses:
        200:
          description: Successful retrieval of settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigData"
        500:
          description: Server error
    put:
      summary: Update configuration settings
      description: Updates the configuration settings
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigData"
      responses:
        200:
          description: Successful update
        500:
          description: Server error
...
components:
  schemas:
...
    ConfigData:
      type: object
      properties:
        BonusServerDomain:
          type: string
          description: 獎勵伺服器域名
        BonusServerPort:
          type: number
          description: 獎勵伺服器端口
        BonusServerPassword:
          type: string
          description: 獎勵伺服器密碼
        UrlUpperDomain:
          type: string
          description: 上層域名
        SubDomains:
          type: string
          description: 子域名
        APITransferPoints:
          type: number
          description: API轉換點數
        CollectSubScale:
          type: number
          description: 收集子域比例
        RabbitMQServer:
          type: string
          description: RabbitMQ伺服器
        RabbitMQUserName:
          type: string
          description: RabbitMQ使用者名稱
        RabbitMQPassword:
          type: string
          description: RabbitMQ密碼
        ConditionWinA:
          type: number
          description: 條件贏金額A
        ConditionWinB:
          type: number
          description: 條件贏金額B
        ConditionWinCR:
          type: number
          description: 條件贏金額CR
...